Added bullet audio, alternating button system
Audio added for when a bullet is created and when it collides with a
player. Altered the player1 and player2 scripts so you move right with a
button press, alternating button scheme. Altered the maxSpeed for
players from 10f to 80f. Multiplying right force by 8 since it is now on
button press.